refactor(host|client): shared HostContext and runtime types, focused vitest
- dsh-tauri host/types: HostContext gains HostLifecycle (effect/inject/
  plugin/logger), shared by extension host/apply (local HostContext removed)
- dsh-tauri/client types/runtime: shared SessionsRuntimeLike /
  WorkspacesRuntimeLike (+ snapshot types); dsh-tauri-session now re-exports
  from dsh-tauri/client instead of declaring its own
- rightclick keeps its official Pick-based runtime faces (reverted)
- vitest.config.ts: run only packages/** tests, maxWorkers 4 + 30s
  timeouts (fixes worktree operation git flake; dsh-core vendored tests
  need @/ paths unavailable here)
