+++ packages/dsh-tauri/src/client/types/runtime.ts
@@ -13,3 +13,71 @@ import type { WorkspaceId as WorkspaceIdType } from './json'
 export type { ISessions, IWorkspaces, SessionListState, SessionSummary, SlotRegistry, WorkspaceSnapshot, WorkspaceSource, WorkspaceView }
 export type SessionsRuntime = ISessions
 export type WorkspacesRuntime = IWorkspaces & { startSession?: (workspaceId?: WorkspaceIdType) => void }
+
+/**
+ * workspace 共享的宽松运行时面（ctx.sessions / ctx.workspaces 投影）。
+ * 消费插件（dsh-tauri-session / dsh-tauri-rightclick / dsh-tauri-pet）统一从
+ * `dsh-tauri/client` 引用，不再各自声明重复面，也不重复 `as unknown as` 断言
+ * （断言目标统一为此处类型）。未知成员用宽松签名，需要官方精确面时用上面的
+ * SessionsRuntime / WorkspacesRuntime Pick。
+ */
+
+/** 会话列表快照（list.subscribe/getSnapshot 投影）。 */
+export interface SessionSummaryLike {
+  id: string
+  title?: string
+  displayTitle?: string
+  cwd?: string
+  updatedAt?: number
+  blank?: boolean
+  /** Session provenance; official sidebar hides subagent sessions. */
+  origin?: string
+}
+
+export interface SessionListSnapshot {
+  ids: string[]
+  byId: Record<string, SessionSummaryLike>
+  current?: string
+  phase?: 'pending' | 'ready'
+}
+
+/** Minimal `ctx.sessions` face the plugins subscribe to. */
+export interface SessionsRuntimeLike {
+  list: {
+    subscribe: (listener: () => void) => () => void
+    getSnapshot: () => SessionListSnapshot
+  }
+  /** Rebuild the in-memory session list after deleting persisted sessions. */
+  refresh?: () => Promise<void>
+  open?: (sessionId: string) => unknown
+  binding?: (sessionId: string) => unknown
+  fork?: (...args: unknown[]) => unknown
+}
+
+/** 工作区视图快照（list 投影）。 */
+export interface WorkspaceViewLike {
+  workspaceId: string
+  path: string
+  title?: string
+  sessionIds: string[]
+}
+
+export interface WorkspaceListSnapshot {
+  items: readonly WorkspaceViewLike[]
+  archivedSessionIds?: readonly string[]
+}
+
+/** Minimal `ctx.workspaces` face the plugins subscribe to. */
+export interface WorkspacesRuntimeLike {
+  list: {
+    subscribe: (listener: () => void) => () => void
+    getSnapshot: () => WorkspaceListSnapshot
+  }
+  /** Wire-truth owner behind the `list` projection. */
+  manager?: {
+    refresh?: () => Promise<void>
+  }
+  archiveSession?: (...args: unknown[]) => unknown
+  delete?: (...args: unknown[]) => unknown
+  startSession?: (workspaceId: string) => unknown
+}
