+++ packages/dsh-tauri/src/host/types/index.ts
@@ -23,5 +23,13 @@ export interface AgentPresetsService { mount?: (agentCtx: unknown, presetId: str
 export interface PermissionPresetsService { names: readonly string[], defaultPreset: string, optionOf: (name: string) => { value: string, name: string, description?: string }, set: (session: unknown, name: string) => void }
 export interface WebServerService { register: (route: HostRoute) => () => void }
 export interface ToolRegistrationService { register: (definition: ToolDefinition | unknown) => () => void | undefined }
-export type HostContext = Context & { agents: AgentRegistry, sessions: SessionStore, workspaceRegistry: WorkspaceRegistry, tools: ToolRuntime & ToolRegistrationService, webServer: WebServerService, connection?: ConnectionGate, llm?: LlmRuntime, agentDefaultModel?: AgentDefaultModelService, agentPresets?: AgentPresetsService, permissionPresets?: PermissionPresetsService, loader: { import: (name: string) => Promise<unknown>, unwrapExports: (exports: unknown) => unknown } }
+export type HostContext = Context & { agents: AgentRegistry, sessions: SessionStore, workspaceRegistry: WorkspaceRegistry, tools: ToolRuntime & ToolRegistrationService, webServer: WebServerService, connection?: ConnectionGate, llm?: LlmRuntime, agentDefaultModel?: AgentDefaultModelService, agentPresets?: AgentPresetsService, permissionPresets?: PermissionPresetsService, loader: { import: (name: string) => Promise<unknown>, unwrapExports: (exports: unknown) => unknown } } & HostLifecycle
+
+/** 宿主生命周期/注册表面（cordis Context 类型发布缺陷兜底：各插件 host/apply 共享）。 */
+export interface HostLifecycle {
+  effect: (callback: () => void | (() => void), id?: string) => void
+  inject: (services: readonly string[], setup: (hostCtx: HostContext) => void) => void
+  plugin: (plugin: unknown, config?: unknown) => unknown
+  logger: { error: (message: string) => void }
+}
 export type { AgentRegistry, ContentBlock, LlmRuntime, ModelSelection, SessionStore, ToolDefinition, ToolRuntime, WorkspaceRegistry }
